Improve resilience of start-logging.ps1
I found I was getting EOFs from my Quest way too often. With this now we keep trying to reconnect the adb logcat "link" until Beat Saber is no longer running or we stop the script.
Special care was given to outputting to a file, where we append to the file on subsequent runs of adb logcat.

Also, increased retries for Beat Saber pid to 5 times with a second of delay.