Implement GetPlayCount and SetAllowCurrentPlayToIncrementCount APIs
See the headers for details on each function.

Currently SetAllowCurrentPlayToIncrementCount is a bit flawed as play counts are incremented as soon as possible, so there isn't really room for other mods to set this flag is e.g. the user has set their progress percentage setting to 0. I have added a TODO item to try and solve this.