Implement custom CanIncrementPlayCountCheck API
- Register and unregister check functions with a name
- Replaces SetAllowCurrentPlayToIncrementCount() API

To solve the problem of SetAllowCurrentPlayToIncrementCount() having a limited time to be called (after gameplay is initiated but before the song progress threshold is passed)
I thought about only incrementing play count when the user finishes a beatmap. But then I noticed I'd be removing functionality just to fix an API quirk.
That is when I noticed the problem was a design flaw on the API, not the quirk itself. So I came up with a better system where other mods can register "can increment play count" checks on load(), and not have to worry about the song process threshold.

Like any good API, I also provide a method to unregister these check functions. Which should be used if a mod is enabled dynamically.